moved connected decorator into hooks.py
